+++ CHANGELOG.md
@@ -9,9 +9,15 @@
 - [依赖库版本升级和安全漏洞修复](https://github.com/buession/buession-parent/releases/tag/v2.3.3)
 
 
+### ⭐ 新特性
+
+- **buession-springboot-boot：** 新增线程池自动配置类 ThreadPoolConfiguration
+
+
 ### ⏪ 优化
 
 - **buession-springboot-pac4j：** 优化 Pac4jFilter 类
+- **buession-springboot-shiro：** 优化 Shiro Filter 初始化及处理流程
 
 
 ---